sensor: lsm6dso: Disable -Wmaybe-uninitialized for lsm6dso_mode_set
This commit disables the potentially uninitialized variable warning
(`-Wmaybe-uninitialized`) for the `lsm6dso_mode_set` function because
GCC 12 and above may report a false positive warning claiming that the
`ctrl2_ois` variable may be used uninitialized.

Note that the `ctrl2_ois` variable is always set when `aux_ctx != NULL`
and is accessed if and only if `aux_ctx != NULL`, therefore it may
never be used uninitialized as the warning claims.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>